Migrating curated storage to ClickHouse (#51)
* adjust everything, worry later

* something working yay

* doing everything everywhere all at once

* testsss

* its aliiiiveeeee

* testss

* cleanups

* more tests, cleanup

* docs fixes

* CR

* Update pipeline-python/README.md


---------

+++ pipeline-python/migrations/clickhouse/init.sql
@@ -4,7 +4,7 @@ CREATE DATABASE IF NOT EXISTS jackfruit;
 -- Each row is a single (variable, timestamp, lat, lon) measurement.
 -- Serves point queries: nearest grid point for a given variable + time.
 --
--- catalog_id references catalog.grid_data.id in Postgres (lineage/metadata).
+-- catalog_id references catalog.curated_data.id in Postgres (lineage/metadata).
 -- The serving layer queries CH for values, then optionally looks up
 -- catalog_id in Postgres for lineage (source, dataset, raw file).
 --
@@ -14,8 +14,8 @@ CREATE DATABASE IF NOT EXISTS jackfruit;
 CREATE TABLE IF NOT EXISTS jackfruit.grid_data (
     variable     LowCardinality(String),
     timestamp    DateTime,
-    lat          Float64,
-    lon          Float64,
+    lat          Float32,
+    lon          Float32,
     value        Float32,
     unit         LowCardinality(String),
     catalog_id   UUID,
